refactor(web): refactor Aspect type definitions and imports
Moved the Aspect interface to a central location in aspectType.ts and updated its structure. Removed duplicate Aspect and AspectTier definitions from page.tsx and aspect-pool-history.tsx, and updated imports and usage to use the unified type.
